test: fix pgTAP bigint overload in harness

+++ supabase/tests/000_foundation_harness.sql
@@ -5,7 +5,7 @@ select plan(2);
 select has_extension('pgtap');
 
 select is(
-  (select count(*) from pg_tables where schemaname = 'public'),
+  (select count(*)::int from pg_tables where schemaname = 'public'),
   0,
   'fresh database starts with no public tables'
 );
